+++ packages/ub-pub/CHANGELOG.md
@@ -6,6 +6,16 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 ### Added
+ - new method `ClientRepository.selectColumunar` - Useful for data analysis \ graph
+```javascript
+await UB.Repository('req_request')
+   .attrs('department.name', 'COUNT([ID])')
+   .groupBy('department.name').select({'department.name': 'department', 'COUNT([ID])': 'requestCount'})
+// {
+//   'department': ["Electricity of Kiev", "Water utility of Kiev"],
+//   'requestCount': [5, 44]
+// }
+```
 
 ### Changed